Fix compiler warnings about misleading indentation, comparisons between signed and unsigned types and an extra semicolon

+++ python_orocos_kdl/PyKDL/pybind11/kinfam.cpp
@@ -227,7 +227,7 @@ void init_kinfam(pybind11::module &m)
     {
         int i = std::get<0>(idx);
         int j = std::get<1>(idx);
-        if (i < 0 || i > 5 || j < 0 || j >= jac.columns())
+        if (i < 0 || i > 5 || j < 0 || (unsigned int)j >= jac.columns())
             throw py::index_error("Jacobian index out of range");
         return jac((unsigned int)i, (unsigned int)j);
     });
@@ -235,7 +235,7 @@ void init_kinfam(pybind11::module &m)
     {
         int i = std::get<0>(idx);
         int j = std::get<1>(idx);
-        if (i < 0 || i > 5 || j < 0 || j >= jac.columns())
+        if (i < 0 || i > 5 || j < 0 || (unsigned int)j >= jac.columns())
             throw py::index_error("Jacobian index out of range");
 
         jac((unsigned int)i, (unsigned int)j) = value;
@@ -265,14 +265,14 @@ void init_kinfam(pybind11::module &m)
     jnt_array.def("resize", &JntArray::resize);
     jnt_array.def("__getitem__", [](const JntArray &ja, int i)
     {
-        if (i < 0 || i >= ja.rows())
+        if (i < 0 || (unsigned int)i >= ja.rows())
             throw py::index_error("JntArray index out of range");
 
         return ja(i);
     });
     jnt_array.def("__setitem__", [](JntArray &ja, int i, double value)
     {
-        if (i < 0 || i >= ja.rows())
+        if (i < 0 || (unsigned int)i >= ja.rows())
             throw py::index_error("JntArray index out of range");
 
         ja(i) = value;
